ui: refactor admin ui
